Block the XML parser on parser blocking stylesheets.
The html spec talks about script blocking stylesheets[1], but we do
script blocking in body by pausing the html parser. The APIs for doing
that for the XML parser had empty implementations which meant that
querying style from a script after an external stylesheet resource would
not wait for the stylesheet to load.

Bug: 1043944, 917116
Change-Id: I08422fa0608db8972154a8b19854906e154dfdd2